+++ src/main/groovy/ru/vyarus/gradle/plugin/mkdocs/MkdocsExtension.groovy
@@ -29,10 +29,10 @@ class MkdocsExtension {
      * {@code python.pip 'mkdocs:0.18.0'} (note that version could be also downgraded).
      */
     static final String[] DEFAULT_MODULES = [
-            'mkdocs:1.0.4',
-            'mkdocs-material:3.0.4',
-            'pygments:2.2.0',
-            'pymdown-extensions:6.0.0',
+            'mkdocs:1.1',
+            'mkdocs-material:4.6.3',
+            'pygments:2.6.1',
+            'pymdown-extensions:6.3.0',
     ]
 
     private final Project project