+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* (breaking) Drop java 7 support
+
 ### 1.1.0 (2018-10-14)
 * Default template's mkdocs.yml changes:
     - fixed `edit_uri` (missed "/docs" ending)